+++ circus-train-core/src/main/java/com/hotels/bdp/circustrain/core/conf/TableParametersConfig.java
@@ -15,16 +15,18 @@
  */
 package com.hotels.bdp.circustrain.core.conf;
 
+import static com.hotels.bdp.circustrain.core.conf.TableParametersConfig.TABLE_REPLICATION_TABLE_PARAMETERS;
+
 import java.util.Map;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ConfigurationProperties(prefix = "table-parameters")
+@ConfigurationProperties(prefix = TABLE_REPLICATION_TABLE_PARAMETERS)
 public class TableParametersConfig {
 
-  public static final String TABLE_REPLICATION_OVERRIDE_TABLE_PARAMETERS = "table-parameters";
+  public static final String TABLE_REPLICATION_TABLE_PARAMETERS = "table-properties";
 
   private Map<String, String> parameters;