+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 ### Removed
 * Excluded `org.pentaho:pentaho-aggdesigner-algorithm` from build.
 
-###Changed
+### Changed
 * Updated `jackson` version to 2.9.9 (was 2.9.8).
+* Updated `circus-train-minimal.yml.template` to include the required `housekeeping` configuration for using the default schema with H2. 
 
 ## [14.0.1] - 2019-04-09
 ### Changed